requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-community
langchain-groq
langchain-chroma
langchain-text-splitters
pypdf
sentence-transformers
chromadb
pydantic[email]
certifi